backend/requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.2
flask-cors==3.0.10
python-dotenv==0.19.0
PyMySQL==1.0.2
flask-jwt-extended==4.3.1
werkzeug==2.3.7
SQLAlchemy==1.4.41
python-jose==3.3.0